Optimizing preloading in library bootstrap.
Adding `Set` as it is always used when accessing the database.

Removing `Apc` file cache adapter preloading as there should not
be a bias upon _which_ cache adapter may be in use.

Removing view related preloading as when using the framework to
just render i.e. json API responses the view part is not involved.