Load routes under console environment too.
Prior to this update, application and plugin routes were not available from the console environment. The filter that loads the routes in bootstrap/action.php, applies to lithium\action\Dispatcher only. The filter logic for loading routes should also be applied to lithium\console\Dispatcher, so that Router related capabilities such as URL generation is possible from the console.